Create Crime Areas detection by clustering
Results and Discussion
The aim of the project would be to everyone who is interested in investing in London have the chance to determine in which Borough would be better to open a restaurant or start a business. So we discuss the best Boroughs so they have success and avoid any possible trouble.
With the following project and objectives we state we get the following results:

We should avoid the following Boroughs:
Westminster
Newham
Southwark
Lambeth
Croydon
We should take in consideration the following Boroughs:
City of London
Kingston upon Thames
Richmond upon Thames
Sutton
Merton
With the folium map proyected they would find out the distance between each other if they decide to live and work in different ones.
Conclusion
To conclude, this project get a better understanding of the principal Boroughs in which it's dangerous and safest to start a restaurant, open a business, or even if you are planning to move there or just have a nice day. Safety is a primary concern for everyone nowadays, so this project would help them to feel secure.